+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/MagicProtection.java
@@ -1,14 +1,17 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
+import insane96mcp.iguanatweaksexpanded.data.generator.ITEDamageTypeTagsProvider;
 import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
 import insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment.protection.ITRProtectionEnchantment;
+import net.minecraft.tags.TagKey;
 import net.minecraft.world.damagesource.DamageSource;
-import net.minecraft.world.damagesource.DamageTypes;
+import net.minecraft.world.damagesource.DamageType;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 
 public class MagicProtection extends ITRProtectionEnchantment {
+    public static TagKey<DamageType> SOURCES_REDUCED = ITEDamageTypeTagsProvider.create("enchantments/protection/magic");
     public MagicProtection() {
         super(Rarity.UNCOMMON);
     }
@@ -42,6 +45,6 @@ public class MagicProtection extends ITRProtectionEnchantment {
     }
 
     public boolean isSourceReduced(DamageSource source) {
-        return source.is(DamageTypes.MAGIC) || source.is(DamageTypes.INDIRECT_MAGIC);
+        return source.is(SOURCES_REDUCED);
     }
 }