+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Fixed soulbound duplicating armor
 * Reduced Soul Steel total Armor (12 -> 11)
+* Added a new Damage Type Tag `iguanatweaksexpanded:enchantments/protection/magic` defining which damage types are reduced by Magic Protection enchantment
 
 ## 1.9.5
 * Reduced Enchanted Cleansed Lapis base table enchantability bonus (5 -> 4)